ml-model/requirements.txt
absl-py==2.1.0
astor==0.8.1
bleach==1.5.0
dnspython>=2.3.0
eventlet>=0.33.3
flask==2.2.5
flask-socketio>=5.3.4
gast==0.6.0
grpcio==1.70.0
html5lib==0.9999999
Markdown==3.7
MarkupSafe==3.0.2
numpy<2
opencv-python==4.11.0.86
protobuf==5.29.3
six==1.17.0
tensorboard==1.8.0
termcolor==2.5.0
torch==2.0.1
torchvision==0.15.2
uvicorn==0.22.0
Werkzeug==3.1.3
wheel==0.45.1
Pillow==9.5.0
